PaymentServiceClient added

+++ src/main/java/com/project/event_management_system/repository/OrganizerRepository.java
@@ -1,9 +1,10 @@
 package com.project.event_management_system.repository;
 
+
 import com.project.event_management_system.model.Organizer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrganizerRepository extends JpaRepository<Organizer, Integer> {
+public interface OrganizerRepository extends JpaRepository<Organizer, Long> {
 }
